feat(custom_auth): jwt 인증 방식 header로 통일
